Move tests loading logic into separate module
